Rearrange and extract functions in bottom bar
