+++ tabs/src/main/java/com/neaniesoft/vermilion/tabs/domain/entities/ActiveTab.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.tabs.domain.entities
+
+sealed class ActiveTab {
+    object None : ActiveTab()
+    object Home : ActiveTab()
+    data class Tab(val id: TabId) : ActiveTab()
+}
